a bunch of stuff

+++ DataClasses/Entities/QuestRequirements/ModifierXQuestRequirement.cs
@@ -7,7 +7,7 @@ public class ModifierXQuestRequirement : QuestRequirement
 {
     public override bool Validate(List<UserEvent> events)
     {
-        if (int.TryParse(Value, out var value))
+        if (double.TryParse(Value, out var value))
         {
             events.RemoveAll(e => e.Modifier < value);
             return events.Any();
